use default sql segment separator
