+++ docs/docs/en/guide/upgrade/incompatible.md
@@ -9,4 +9,4 @@ This document records the incompatible updates between each version. You need to
 ## 3.0.0
 
 * Copy and import workflow without 'copy' suffix [#10607](https://github.com/apache/dolphinscheduler/pull/10607)
-
+* Use semicolon as default sql segment separator [#10869](https://github.com/apache/dolphinscheduler/pull/10869)
