+++ docs/docs/zh/guide/task/sql.md
@@ -20,9 +20,7 @@ SQL任务类型，用于连接数据库并执行相应SQL。
 - sql类型：支持查询和非查询两种。
   - 查询：支持 `DML select` 类型的命令，是有结果集返回的，可以指定邮件通知为表格、附件或表格附件三种模板；
   - 非查询：支持 `DDL`全部命令 和 `DML update、delete、insert` 三种类型的命令；
-    - 分段执行符号：提供在数据源不支持一次执行多段SQL语句时，拆分SQL语句的符号来进行多次调用数据源执行方法。
-      例子：1.当数据源选择Hive数据源时，不需要填写此参数。因为Hive数据源本身支持一次执行多段SQL语句；
-      2.当数据源选择MySQL数据源时，并且要执行多段SQL语句时，需要填写此参数为分号 `;`。因为MySQL数据源不支持一次执行多段SQL语句；
+    - 默认采用`;\n`作为SQL分隔符,拆分成多段SQL语句执行,Hive支持一次执行多段SQL语句,故不会拆分。
 - sql参数：输入参数格式为key1=value1;key2=value2…
 - sql语句：SQL语句
 - UDF函数：对于HIVE类型的数据源，可以引用资源中心中创建的UDF函数，其他类型的数据源暂不支持UDF函数。
