+++ dolphinscheduler-task-plugin/dolphinscheduler-task-api/src/main/java/org/apache/dolphinscheduler/plugin/task/api/parameters/SqlParameters.java
@@ -114,15 +114,6 @@ public class SqlParameters extends AbstractParameters {
 
     private int limit;
 
-    /**
-     * segment separator
-     *
-     * <p>The segment separator is used
-     * when the data source does not support multi-segment SQL execution,
-     * and the client needs to split the SQL and execute it multiple times.</p>
-     */
-    private String segmentSeparator;
-
     public int getLimit() {
         return limit;
     }
@@ -235,14 +226,6 @@ public class SqlParameters extends AbstractParameters {
         this.groupId = groupId;
     }
 
-    public String getSegmentSeparator() {
-        return segmentSeparator;
-    }
-
-    public void setSegmentSeparator(String segmentSeparator) {
-        this.segmentSeparator = segmentSeparator;
-    }
-
     @Override
     public boolean checkParameters() {
         return datasource != 0 && StringUtils.isNotEmpty(type) && StringUtils.isNotEmpty(sql);
@@ -310,7 +293,6 @@ public class SqlParameters extends AbstractParameters {
                 + ", sendEmail=" + sendEmail
                 + ", displayRows=" + displayRows
                 + ", limit=" + limit
-                + ", segmentSeparator=" + segmentSeparator
                 + ", udfs='" + udfs + '\''
                 + ", showType='" + showType + '\''
                 + ", connParams='" + connParams + '\''
