+++ dolphinscheduler-ui/src/views/projects/task/components/node/fields/use-sql-type.ts
@@ -24,7 +24,6 @@ import type { IJsonItem } from '../types'
 export function useSqlType(model: { [field: string]: any }): IJsonItem[] {
   const { t } = useI18n()
   const querySpan = computed(() => (model.sqlType === '0' ? 6 : 0))
-  const nonQuerySpan = computed(() => (model.sqlType === '1' ? 18 : 0))
   const emailSpan = computed(() =>
     model.sqlType === '0' && model.sendEmail ? 24 : 0
   )
@@ -67,15 +66,6 @@ export function useSqlType(model: { [field: string]: any }): IJsonItem[] {
         trigger: ['input', 'blur'],
         required: true
       }
-    },
-    {
-      type: 'input',
-      field: 'segmentSeparator',
-      name: t('project.node.segment_separator'),
-      props: {
-        placeholder: t('project.node.segment_separator_tips')
-      },
-      span: nonQuerySpan
     },
     {
       type: 'switch',
